Adding a claim nr sequence for recordings from each First Nation
Enables automatic numbering of ALUR recordings (opentenure.claim)
that based on a First Nation alpha code (new field in opentenure.
administrative_boundary table) and a sequention 4 digit sequential
number. Sequences are created using PG Functions that are called
when a First Nation (administrative_boundary) is saved for the first
time. Replaces previous generic numbering sequence for all recordings.
As previously sequences are called when the claim (recording is saved
for the first time.
Tested on Payara Server 5 Community Edition 5.2021.3 and PostgreSQL
11 & 12.